# docs/assets/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Health Check","operationId":"health_check__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/applications/{app_instance}":{"put":{"summary":"Update Application","operationId":"update_application_applications__app_instance__put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"app_instance","in":"path","required":true,"schema":{"type":"string","title":"App Instance"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"requestBody":{"content":{"application/x-yaml":{"schema":{"$defs":{"AppComponent":{"properties":{"name":{"title":"Name","type":"string"}},"required":["name"],"title":"AppComponent","type":"object"}},"properties":{"name":{"title":"Name","type":"string"},"components":{"items":{"$ref":"#/$defs/AppComponent"},"title":"Components","type":"array"}},"required":["name","components"],"title":"AppDescriptor","type":"object"}}},"required":true}},"delete":{"summary":"Delete Application","operationId":"delete_application_applications__app_instance__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"app_instance","in":"path","required":true,"schema":{"type":"string","title":"App Instance"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/applications":{"get":{"summary":"List Applications","operationId":"list_applications_applications_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}